fix migration script
